Modify: adjust the logic of invitePlayerJoinGame
FYI: Before inserting data, first determine whether there is a combination of the same invite_id-game_id in the db. If not, insert it.
FYI: If you receive an invitation, the association between the player and the game will be established directly; if you do not enter the public game through an invitation, the association between the player and the game will be established when the first puzzle is correctly locked in game by you; for players who have not signined in, it will not have any connection with the game

1. Feat: add create_at field to player_game table for easier reference
2. Modify: pass in an additional playerId attribute when sending updateAndLockPiece socket event